requirements.txt
flask
flask-cors
requests
bs4
apscheduler
pymongo
python-dotenv